Refactor src directory
