+++ cron.sh
@@ -4,5 +4,5 @@
 echo "Running cron job at `date`" >> /app/cron.log
 
 # Run the python script
-python3 -m flask --app train_app save-ppm-data
-python3 -m flask --app train_app prune-data
+python3 -m flask --app rail_rank save-ppm-data
+python3 -m flask --app rail_rank prune-data
