+++ pyproject.toml
@@ -54,5 +54,5 @@ ignore_missing_imports = true
 dialect = "sqlite"
 
 [project]
-name = "flask-spark_template"
+name = "rail_rank"
 requires-python = ">=3.10"
