+++ rail_rank/static/bulma.css
