+++ rail_rank/templates/base.html
