+++ rail_rank/templates/rankings/index.html
